Start of implementing use OpenWeatherMap API for forecast

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/domain/model/base/openWeatherMap/IOWMWind.kt
@@ -6,9 +6,9 @@ interface IOWMWind {
     /**Wind speed.
      *
      *Unit Default: meter/sec, Metric: meter/sec, Imperial: miles/hour.*/
-    val speedInUnits: Int
+    val speedInUnits: Double
 
     /**Wind direction, degrees (meteorological)*/
-    val directionInDegrees: Int
+    val directionInDegrees: Double
 
 }
