+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/domain/model/implementation/openWeatherMap/OWMWind.kt
@@ -4,8 +4,8 @@ import ru.mts.avpopo85.weathery.domain.model.base.openWeatherMap.IOWMWind
 
 data class OWMWind(
 
-    override val speedInUnits: Int,
+    override val speedInUnits: Double,
 
-    override val directionInDegrees: Int
+    override val directionInDegrees: Double
 
 ) : IOWMWind
